# backend/api/swagger.yaml
openapi: 3.0.3
info:
  title: Fantasm API
  version: 2.0.0
  description: |
    JSON API of the Fantasm backend. Every route lives under `/api`.

    **Authentication.** Sign-in is OIDC only (Google, Microsoft Entra). A successful callback sets an
    opaque, HttpOnly, SameSite=Lax session cookie (`__Host-fantasm_session` over HTTPS). There are no passwords.

    **CSRF.** Unsafe methods (POST, PUT, PATCH, DELETE) must come from the site's own origin: a request with a
    foreign `Origin` or `Sec-Fetch-Site: cross-site` is answered with 403.

    **Errors.** Every failure has the same body, see `Error`, including 404 for an unknown path and 405 (with `Allow`) for a wrong method. Hidden, deleted and nonexistent content all
    answer 404 identically, so existence is never revealed. Unexpected failures are a generic 500 with a `requestId`.

    **Pagination.** Lists are keyset-paginated: pass the previous page's `nextCursor` as `cursor`. Cursors are
    signed and bound to the listing they came from. `limit` is clamped to the documented maximum.

    **Limits.** Request bodies are capped (1 MiB), unknown JSON fields are rejected (400), and every route is
    rate limited per user (signed in) or per address (anonymous). Limited answers carry `Retry-After`.
servers:
  - url: /
security: []
...
paths:
...
  /api/ideas:
    get:
      tags: [ideas]
      summary: Feed of public ideas
      parameters:
        - { name: sort, in: query, schema: { type: string, enum: [hot, new, top], default: hot } }
        - { name: category, in: query, schema: { $ref: '#/components/schemas/Category' } }
        - { name: status, in: query, schema: { $ref: '#/components/schemas/Status' } }
        - { name: tag, in: query, schema: { type: string } }
        - { name: campus, in: query, schema: { type: string } }
        - $ref: '#/components/parameters/Cursor'
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 50, default: 20 } }
      responses:
        '200':
          description: One page of idea cards.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/IdeaPage' }
        '422': { $ref: '#/components/responses/Validation' }
    post:
      tags: [ideas]
      summary: Create an idea
      description: |
        Authors with at least three approved ideas, or an organization benefit that skips premoderation, are
        approved at once; everyone else enters the moderation queue as PENDING. Drafts are not moderated until
        published. At most five ideas a day for users without that benefit.
      security: [{ session: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/IdeaCreate' }
      responses:
        '201':
          description: Created.
          headers:
            Location: { schema: { type: string } }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/IdeaView' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '422': { $ref: '#/components/responses/Validation' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /api/ideas/{slug}:
    parameters:
      - $ref: '#/components/parameters/Slug'
    get:
      tags: [ideas]
      summary: One idea
      description: Authors see their own unpublished or hidden ideas and staff see everything; for everybody else those answer 404.
      responses:
        '200':
          description: The idea, rendered and sanitized.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/IdeaView' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [ideas]
      summary: Edit an idea (author, or staff)
      security: [{ session: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/IdeaPatch' }
      responses:
        '200':
          description: Updated idea.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/IdeaView' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/Validation' }
    delete:
      tags: [ideas]
      summary: Delete an idea (soft delete)
      security: [{ session: [] }]
      responses:
        '204': { description: "Deleted." }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/ideas/{slug}/status:
    post:
      tags: [ideas]
      summary: Move an idea through its lifecycle
      description: Publishing a DRAFT sends it to premoderation unless the author is trusted.
      security: [{ session: [] }]
      parameters:
        - $ref: '#/components/parameters/Slug'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [status]
              properties:
                status: { $ref: '#/components/schemas/Status' }
      responses:
        '200':
          description: The idea after the change.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/IdeaView' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/Validation' }
  /api/ideas/{slug}/next:
    get:
      tags: [ideas]
      summary: Next idea to read
      parameters:
        - $ref: '#/components/parameters/Slug'
      responses:
        '200':
          description: A short reference to the next idea in the feed.
          content:
            application/json:
              schema:
                type: object
                properties:
                  slug: { type: string }
                  title: { type: string }
                  summary: { type: string }
                  category: { $ref: '#/components/schemas/Category' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/events:
    get:
      tags: [ideas]
      summary: Upcoming events
      parameters:
        - { name: from, in: query, schema: { type: string, format: date-time } }
        - { name: to, in: query, schema: { type: string, format: date-time } }
      responses:
        '200':
          description: Event ideas the viewer may see, soonest first.
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/IdeaCard' }
        '422': { $ref: '#/components/responses/Validation' }
...
  /api/me/ideas:
    get:
      tags: [ideas]
      summary: Own ideas, including drafts and moderation state
      security: [{ session: [] }]
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 50, default: 20 } }
      responses:
        '200':
          description: One page.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/IdeaPage' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/users/{handle}:
    get:
      tags: [ideas]
      summary: Public profile with the author's visible ideas
      parameters:
        - $ref: '#/components/parameters/Handle'
      responses:
        '200':
          description: Profile.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PublicProfile' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: __Host-fantasm_session
      description: Opaque session token set by the OIDC callback. Only its SHA-256 hash is stored.
  parameters:
...
    Slug:
      name: slug
      in: path
      required: true
      schema: { type: string, pattern: '^[a-z0-9]+(-[a-z0-9]+)*$', maxLength: 90 }
    Handle:
      name: handle
      in: path
      required: true
      schema: { type: string, pattern: '^[a-z0-9_][a-z0-9_-]{2,39}$' }
    Cursor:
      name: cursor
      in: query
      description: The `nextCursor` of the previous page. A forged or foreign cursor is a 422.
      schema: { type: string, maxLength: 512 }
  responses:
    Unauthorized:
      description: Not signed in, or the session expired.
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    Forbidden:
      description: Signed in but not allowed, or a cross-origin request.
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    NotFound:
      description: No such resource, or one the viewer may not know about.
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
...
    Validation:
      description: The input is well-formed but not acceptable; `fields` names what to fix.
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    RateLimited:
      description: Too many requests.
      headers:
        Retry-After: { schema: { type: integer }, description: Seconds to wait. }
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
  schemas:
    Error:
      type: object
      required: [error, code]
      properties:
        error: { type: string }
        code:
          type: string
          enum: [invalid_request, unauthorized, forbidden, not_found, conflict, rate_limited, payload_too_large, validation_failed, unsupported_media_type, method_not_allowed, not_implemented, internal_error]
        fields: { type: object, additionalProperties: { type: string } }
        requestId: { type: string }
    Category: { type: string, enum: [STARTUP, PROJECT, EVENT, COMMUNITY, VOLUNTEERING, OTHER] }
    Status: { type: string, enum: [DRAFT, OPEN, TEAM_FORMING, IN_PROGRESS, DONE, ARCHIVED] }
    Visibility: { type: string, enum: [PUBLIC, MEMBERS_ONLY, ORGANIZATION_ONLY] }
    Campus:
      type: object
      nullable: true
      properties:
        id: { type: string }
        label: { type: string }
...
    Tag:
      type: object
      properties:
        slug: { type: string }
        label: { type: string }
    IdeaAuthor:
      type: object
      properties:
        handle: { type: string }
        name: { type: string }
        verified: { type: boolean }
        campus: { $ref: '#/components/schemas/Campus' }
    IdeaCard:
      type: object
      properties:
        slug: { type: string }
        title: { type: string }
        summary: { type: string }
        story: { type: string, description: "Sanitized plain-text excerpt for cards." }
        coverUrl: { type: string }
        category: { $ref: '#/components/schemas/Category' }
        status: { $ref: '#/components/schemas/Status' }
        moderation:
          type: string
          enum: [PENDING, APPROVED, HIDDEN, REJECTED]
          description: Only shown to the author and staff.
        visibility: { $ref: '#/components/schemas/Visibility' }
        organizationId: { type: string, nullable: true }
        campus: { $ref: '#/components/schemas/Campus' }
        tags: { type: array, items: { $ref: '#/components/schemas/Tag' } }
        author: { $ref: '#/components/schemas/IdeaAuthor' }
        votes: { type: integer, description: "Weighted votes." }
        comments: { type: integer }
        participants: { type: integer }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        publishedAt: { type: string, format: date-time, nullable: true }
        eventAt: { type: string, format: date-time }
        eventLocation: { type: string }
        eventCapacity: { type: integer }
        needsRoles: { type: array, items: { type: string } }
    IdeaView:
      allOf:
        - $ref: '#/components/schemas/IdeaCard'
        - type: object
          properties:
            html: { type: string, description: "Sanitized HTML, safe to inject." }
            text: { type: string }
            toc:
              type: array
              items:
                type: object
                properties:
                  id: { type: string }
                  text: { type: string }
                  level: { type: integer }
            readingMinutes: { type: integer }
            body: { type: string, description: "Markdown source. Only for those who can edit." }
            canEdit: { type: boolean }
            viewer:
              type: object
              properties:
                voted: { type: boolean }
                votedWeight: { type: integer }
                participation:
                  type: object
                  nullable: true
                  properties:
                    state: { type: string, enum: [INTERESTED, JOINED, ACCEPTED, DECLINED] }
                    role: { type: string }
    IdeaPage:
      type: object
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/IdeaCard' } }
        nextCursor: { type: string, nullable: true }
    IdeaCreate:
      type: object
      additionalProperties: false
      required: [title, category]
      properties:
        title: { type: string, minLength: 3, maxLength: 120 }
        summary: { type: string, maxLength: 280 }
        body: { type: string, maxLength: 20000, description: "Markdown. Raw HTML is dropped, images must be https." }
        coverUrl: { type: string, format: uri, maxLength: 500, description: "https only." }
        category: { $ref: '#/components/schemas/Category' }
        tags: { type: array, maxItems: 5, items: { type: string, maxLength: 40 } }
        visibility: { $ref: '#/components/schemas/Visibility' }
        organizationId: { type: string }
        status: { type: string, enum: [DRAFT, OPEN] }
        eventAt: { type: string, format: date-time }
        eventLocation: { type: string, maxLength: 200 }
        eventCapacity: { type: integer, minimum: 1, maximum: 100000 }
        needsRoles: { type: array, maxItems: 10, items: { type: string, maxLength: 40 } }
    IdeaPatch:
      type: object
      additionalProperties: false
      properties:
        title: { type: string }
        summary: { type: string }
        body: { type: string }
        coverUrl: { type: string }
        category: { $ref: '#/components/schemas/Category' }
        tags: { type: array, items: { type: string } }
        eventAt: { type: string, format: date-time }
        eventLocation: { type: string }
        eventCapacity: { type: integer }
        needsRoles: { type: array, items: { type: string } }
    PublicProfile:
      type: object
      properties:
        handle: { type: string }
        name: { type: string }
        bio: { type: string }
        faculty: { type: string }
        verified: { type: boolean }
        campus: { $ref: '#/components/schemas/Campus' }
        role: { type: string, enum: [USER, MODERATOR, ADMIN] }
        karma: { type: integer }
        joinedAt: { type: string, format: date-time }
        ideasCount: { type: integer }
        ideas: { type: array, items: { $ref: '#/components/schemas/IdeaCard' } }
...